refactor : checkout details
